Set up outline for amqp and mqtt consumers
