+++ container_service_extension/mqtt_consumer.py
@@ -1,0 +1,3 @@
+class MQTTConsumer:
+    def __init__(self):
+        pass
